+++ subs/rio/src/RIO.hs
@@ -1,0 +1,4 @@
+module RIO (module X) where
+
+import RIO.Prelude as X
+import RIO.Logger as X
