--- subs/rio/src/System/Process/Log.hs
@@ -1,55 +1,0 @@
-{-# LANGUAGE NoImplicitPrelude #-}
-{-# LANGUAGE OverloadedStrings #-}
-
--- | Separate module because TH.
-
-module System.Process.Log
-    (withProcessTimeLog -- FIXME move into System.Process.Read
-    ,showProcessArgDebug)
-    where
-
-import qualified Data.Text as T
-import           Stack.Prelude
-import qualified System.Clock as Clock
-
--- | Log running a process with its arguments, for debugging (-v).
---
--- This logs one message before running the process and one message after.
-withProcessTimeLog :: (MonadIO m, MonadReader env m, HasLogFunc env, HasCallStack) => Maybe FilePath -> String -> [String] -> m a -> m a
-withProcessTimeLog mdir name args proc = do
-  let cmdText =
-          T.intercalate
-              " "
-              (T.pack name : map showProcessArgDebug args)
-      dirMsg =
-        case mdir of
-          Nothing -> ""
-          Just dir -> " within " <> T.pack dir
-  logDebug ("Run process" <> dirMsg <> ": " <> cmdText)
-  start <- liftIO $ Clock.getTime Clock.Monotonic
-  x <- proc
-  end <- liftIO $ Clock.getTime Clock.Monotonic
-  let diff = Clock.diffTimeSpec start end
-  -- useAnsi <- asks getAnsiTerminal
-  let useAnsi = True
-  logDebug
-      ("Process finished in " <>
-      (if useAnsi then "\ESC[92m" else "") <> -- green
-      timeSpecMilliSecondText diff <>
-      (if useAnsi then "\ESC[0m" else "") <> -- reset
-       ": " <> cmdText)
-  return x
-
-timeSpecMilliSecondText :: Clock.TimeSpec -> Text
-timeSpecMilliSecondText t =
-    (T.pack . show . (`div` 10^(6 :: Int)) . Clock.toNanoSecs) t <> "ms"
-
--- | Show a process arg including speechmarks when necessary. Just for
--- debugging purposes, not functionally important.
-showProcessArgDebug :: String -> Text
-showProcessArgDebug x
-    | any special x || null x = T.pack (show x)
-    | otherwise = T.pack x
-  where special '"' = True
-        special ' ' = True
-        special _ = False
